#### Reference Issues/PRs
Could be seen part of #20132 or follow-up of #22206.

#### What does this implement/fix? Explain your changes.
Currently in the `QuantileRegressor`, `X` is used twice to construct a sparse matrix `A_eq`. This PR introduces "new" variables such that `X` enters `A_eq` only once. The price are `n_features` additional rows in `A_eq`, but they are very sparse.
This saves memory and it seems to have no noticeable impact on fit time.

#### Any other comments?
This was originally proposed in https://github.com/scipy/scipy/issues/15185#issuecomment-991716877.

My main reason to open this PR is that the knowledge is not lost.