@lorentzenchr 2022-01-15
For core devs: Do we have a github label for memory footprint? Do we use the "performance" label for this? Or should we create a new one?