@thomasjpfan 2022-01-15
> Do we have a github label for memory footprint? Do we use the "performance" label for this?

Yes it's the "Performance" label and in this case, "memory performance". 

> Or should we create a new one?

I'm happy with the generic "Performance" label for now.