@ogrisel 2022-01-26
Unfortunately I observe a negative impact in terms of CPU time:

## On main

```python
In [6]: %timeit QuantileRegressor(solver="highs").fit(*make_classification(n_samples=10_000, n_features=10, random_state=0))
1.24 s ± 6.43 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

## On this PR

```python
In [4]: %timeit QuantileRegressor(solver="highs").fit(*make_classification(n_samples=10_000, n_features=10, random_state=0))
3.15 s ± 89.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Maybe we could introduce an option to switch between the 2 formulations to trade memory usage for CPU speed?

I have the feeling that given the relative slowness of this optimization problem, the CPU time will be more of a problem than the memory usage for most users.

Building a problem where X is larger that a few GB is problem untractable from a CPU time point of view.