@ogrisel 2022-01-26
I am fine with using the Performance label for this issue because it impacts both memory and CPU performance. But feel free to introduce a dedicated label if you prefer. (e.g. "Memory Usage").