@lorentzenchr 2023-03-01
> To you, what is the status of this PR? Should we pursue it?

I guess we can close.